Fix error in test
If the ego agent is one of the random agents, the goal polygon of the
agent has already been set. By redefining it using the goal position
defined in the parameters, it will no longer be located in the same lane
as the ego agent, leading to an incorrectly generated DrivingCorridor.